fix(packaging): one clock for lite builds too, decode reg output correctly, test the masker
Review round 7 on #49.

- composeApp/build.gradle.kts documented the iteration I abandoned (run_number default,
  callers passing build_id, "disjoint range"). It is the file release.yml points at, so it
  now describes what actually ships: prepare-version computes one seconds-since-epoch stamp
  per run, the macOS job reads it from `needs`, and every trigger resolves the same
  expression.
- release-lite.yml still stamped github.run_number, and BOSS Lite ships the same bundle id
  (ai.rever.boss) — so one bundle id had two independent CFBundleVersion sequences: lite
  builds sorting permanently below every direct release, and lite run N able to stamp a value
  release.yml run N already shipped. It now uses the same clock, computed in its own
  prepare-version job.
- reg.exe writes in the console/OEM code page, and the redirected output was decoded with the
  JVM default (UTF-8 on 18+). For an install path like C:\Users\Björn\... that yields U+FFFD,
  which parses fine, is not our path, and does not exist on disk — so the classifier called a
  live third-party registration dead and deleted it. Output is now decoded with
  native.encoding, and the pure layer treats any path containing U+FFFD as UNREADABLE, the
  same "not evaluatable ⇒ leave alone" rule as %VARIABLE%.
- maskUserPath's %USERPROFILE% replacement replaced the string with itself. It was also the
  one new pure function left untested and outside the policy file — which is how a no-op
  survived in a PR arguing "the policy is pure, so test it". Dropped the dead line, moved the
  function into the policy object, and added assertions for both real forms and null.
- The policy object is now WindowsProtocolCleanup in WindowsProtocolCleanup.kt: the generic
  internal names (parseCommandState, exitCodeFor, ...) are namespaced instead of sitting loose
  in ai.rever.boss.utils.
- Smaller review notes: deleteOnExit() backstop on the temp file, since destroyForcibly is
  asynchronous and Windows may still hold the handle when `finally` deletes; the two remaining
  per-call regexes hoisted to vals alongside UNEXPANDED_VARIABLE; the date +%s granularity
  caveats (same-second approvals, re-run reusing the cached output) written into the comment;
  and three test gaps closed — unknown root key outranking a readable command, both delete
  predicates true (the ordinary uninstall shape), and the mis-decode regression.

34 tests in WindowsProtocolCleanupTest, 18 in DebControlTest.
