+++ .github/workflows/release.yml
@@ -112,6 +112,12 @@ jobs:
         # Single non-negative integer, so Compose's CFBundleVersion validation accepts it;
         # fits a signed 32-bit int until 2038. CI only — local builds use app.build.number,
         # which is stable so packaging tasks stay up-to-date between runs.
+        #
+        # Granularity caveats, neither of which can collide with a *shipped* build: two runs
+        # approved in the same second would share an id (there is no concurrency group on
+        # this workflow), and "Re-run failed jobs" reuses this job's cached output, so a
+        # re-run macOS job stamps the original attempt's id — which is what you want, since
+        # it is the same release.
         run: echo "build-id=$(date +%s)" >> "$GITHUB_OUTPUT"
       - name: 📥 Checkout Repository
         uses: actions/checkout@v7
